### 2024-04-29 #6691 @editorialbot
[PRE REVIEW]: primerForge: a Python program for identifying primer pairs capable of distinguishing groups of genomes from each other
**Submitting author:** @dr-joe-wirth (<a href="http://orcid.org/0000-0002-9750-2845">Joseph S. Wirth</a>)
**Repository:** https://github.com/dr-joe-wirth/primerForge/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
