### 2024-05-07 #6691 @dr-joe-wirth
Hello @jromanowska and @luizirber,

This is my first submission to JOSS so I am not entirely sure how this process works but based on the list that `@editorialbot` gave me, it seems like both of you may be suitable to serve as an editor for my submission. Thanks for your help and sorry if I was not supposed to tag you like this.
