### 2024-05-08 #6691 @jromanowska
@dr-joe-wirth Thank you for the submission. This is not your task to find reviewers. Please wait for one of the Editors in Chief in JOSS to take care of the first steps and to find an editor for you, who will guide you through the submission process.
I am not able to help you at this time.
