### 2024-05-15 ffa6dba
moved `del` statements into appropriate scope
