### 2024-05-16 1422afb
major overhaul; removed plotting and analysis
