+++ bin/unit_tests/analysisData_test.py
@@ -1,373 +1,0 @@
-from __future__ import annotations
-import unittest
-from bin.Primer import Primer
-from itertools import combinations
-from bin.AnalysisData import AnalysisData, Level
-
-
-class AnalysisDataTest(unittest.TestCase):
-    # global attributes
-    LEVEL_STR = ('candidate kmer', 'unfiltered pair', 'filtered pair', 'final pair')
-    PRIMER = Primer('atcg', 'contig', 8, 4, Primer.PLUS)
-    INDEX = 0
-    NAME = 'genome' 
-    
-    def setUp(self) -> None:
-        return super().setUp()
-    
-    def tearDown(self) -> None:
-        return super().tearDown()
-    
-    def testA_constructor(self) -> None:
-        """make sure attributes have expected values
-        """
-        data = AnalysisData(AnalysisDataTest.PRIMER, AnalysisDataTest.INDEX, AnalysisDataTest.NAME)
-        self.assertEqual(data.primer, self.PRIMER)
-        self.assertEqual(data.getIndex(), self.INDEX)
-        self.assertEqual(data.name, self.NAME)
-        
-    def testB_setGetLevel(self) -> None:
-        """checks that Level.setLevel, AnalysisData.setLevel, and AnalysisData.getLevel are working
-        """
-        # initialize a Level object
-        level = Level()
-        
-        # make sure the constructor worked
-        self.assertEqual(str(level), AnalysisDataTest.LEVEL_STR[0])
-        
-        # check that setLevel and str works for all levels
-        for lev in AnalysisDataTest.LEVEL_STR:
-            # make sure setLevel works with a string
-            level.setLevel(lev)
-            self.assertEqual(str(level), lev)
-            
-            # make sure setLevel works with a Level
-            newLev = Level()
-            newLev.setLevel(lev)
-            level.setLevel(newLev)
-            self.assertEqual(str(level), lev)
-        
-        # check that setLevel fails with invalid inputs
-        self.assertRaises(ValueError, level.setLevel, 'bad level')
-        self.assertRaises(ValueError, level.setLevel, 0)
-
-        # initialize an AnalysisData object
-        data = AnalysisData(AnalysisDataTest.PRIMER, AnalysisDataTest.INDEX, AnalysisDataTest.NAME)
-        
-        # check that setLevel and getLevel works for all levels
-        for lev in AnalysisDataTest.LEVEL_STR:
-            # make sure setLevel works with a string
-            data.setLevel(lev)
-            self.assertEqual(data.getLevel(), lev)
-            
-            # make sure setLevel works with a Level
-            newLev = Level()
-            newLev.setLevel(lev)
-            level.setLevel(newLev)
-            self.assertEqual(data.getLevel(), lev)
-        
-        # check that setLevel fails with invalid inputs
-        self.assertRaises(ValueError, data.setLevel, 'bad level')
-        self.assertRaises(ValueError, data.setLevel, 0)
-    
-    def testC_LevelEquality(self) -> None:
-        """checks that the == operator works for Level class
-        """
-        # initialize levels
-        level = Level()
-        other = Level()
-        
-        # they should be equal
-        self.assertEqual(level, other)
-        
-        # make sure they stay equal for each level
-        for lev in AnalysisDataTest.LEVEL_STR:
-            level.setLevel(lev)
-            other.setLevel(lev)
-            
-            # should work for levels
-            self.assertEqual(level, other)
-            
-            # should work for strings
-            self.assertEqual(level, lev)
-        
-        # make sure it fails with int
-        self.assertRaises(TypeError, level.__eq__, 1)
-        
-    def testD_LevelInequality(self) -> None:
-        """checks that != operator works for Level class
-        """
[…]
