+++ bin/main.py
@@ -5,11 +5,9 @@ from bin.Clock import Clock
 from bin.Primer import Primer
 from Bio.SeqRecord import SeqRecord
 from bin.Parameters import Parameters
-from bin.AnalysisData import AnalysisData
 from bin.getCandidateKmers import _getAllCandidateKmers
 from bin.removeOutgroupPrimers import _removeOutgroupPrimers
 from bin.getPrimerPairs import _getPrimerPairs, _keepOnePairPerBinPair
-from bin.analysis import _initializeAnalysisData, _updateAnalysisData, _plotAnalysisData
 
 # global constants
 __version__ = "1.0.2"
@@ -19,17 +17,16 @@ __CAND_NUM   = 1
 __PAIR_1_NUM = 2
 __PAIR_2_NUM = 3
 __PAIR_3_NUM = 4
-__ANAL_NUM   = 5
 
 
-def __getCheckpoint(params:Parameters) -> tuple[bool,bool,bool,bool,bool,bool]:
+def __getCheckpoint(params:Parameters) -> tuple[bool,bool,bool,bool,bool]:
     """gets booleans for checkpointing
 
     Args:
         params (Parameters): a Parameters object
 
     Returns:
-        tuple[bool,bool,bool,bool,bool,bool]: each boolean indicates if the pickle exists: shared kmers; candidate kmers; unfiltered pairs; filtered pairs; final pairs; analysis data
+        tuple[bool,bool,bool,bool,bool,bool]: each boolean indicates if the pickle exists: shared kmers; candidate kmers; unfiltered pairs; filtered pairs; final pairs
     """
     
     # determine if each pickle exists
@@ -38,9 +35,8 @@ def __getCheckpoint(params:Parameters) -> tuple[bool,bool,bool,bool,bool,bool]:
     unfiltExists = os.path.exists(params.pickles[__PAIR_1_NUM])
     filterExists = os.path.exists(params.pickles[__PAIR_2_NUM])
     finalsExists = os.path.exists(params.pickles[__PAIR_3_NUM])
-    analysExists = os.path.exists(params.pickles[__ANAL_NUM])
     
-    return sharedExists, candidExists, unfiltExists, filterExists, finalsExists, analysExists
+    return sharedExists, candidExists, unfiltExists, filterExists, finalsExists
 
 
 def __readSequenceData(seqFiles:list[str], frmt:str) -> dict[str, list[SeqRecord]]:
@@ -104,48 +100,12 @@ def __getCandidates(params:Parameters, sharedExists:bool, clock:Clock) -> dict[s
     return candidateKmers
 
 
-def __initializeAnalysis(params:Parameters, candidateKmers:dict[str,dict[str,list[Primer]]], clock:Clock) -> dict[tuple[Seq,str],AnalysisData]:
-    """initialize the analysis data
-
-    Args:
-        params (Parameters): a Parameters object
-        candidateKmers (dict[str,dict[str,list[Primer]]]): the dictionary produced by __getCandidates
-        clock (Clock): a Clock object
-
-    Returns:
-        dict[tuple[Seq,str],AnalysisData]: key=(kmer sequence, contig name); val=AnalysisData object
-    """
-    
-    # message
-    MSG  = "initializing analysis data"
-    
-    # log data
-    params.log.rename(__initializeAnalysis.__name__)
-    
-    # print status
-    params.log.info(MSG)
-    clock.printStart(MSG)
-    
-    # initialize analysis data
-    analysisData = _initializeAnalysisData(candidateKmers)
-    
-    # print status
-    clock.printDone()
-    params.log.info(f"done {clock.getTimeString()}")
-    
-    # dump the analysis to file
-    params.dumpObj(analysisData, params.pickles[__ANAL_NUM], "initial analysis data")
-    
-    return analysisData
-
-
[…]

[…]
