### 2024-05-16 3a1ddec
parallelized kmer retrieval
