### 2024-05-16 94dd6fd
revised `Clock._killWheel`
