### 2024-05-21 a6940a3
improved print statements in `__getFinalPairs`
