### 2024-05-21 842a8e3
candidate primer pairs now only retrieved for one genome
