### 2024-05-21 f123d79
improved print for `__writePrimerPairs`
