### 2024-05-22 1d9f1f0
fixed bug in `__updateBinsForUnprocessedGenomes`
