### 2024-05-23 db2dc19
parallelized kmer retrieval
