### 2024-05-23 c1ce512
fixed typo in help message
