### 2024-05-30 b8d92eb
removed `matplotlib` requirement
