+++ environment.yml
@@ -13,7 +13,6 @@ dependencies:
   - python=3.11.0
   - pip=23.0.1
   - biopython=1.81
-  - matplotlib=3.7.2
   - numpy
   - primer3-py>=2.0
   - scipy>=1.10
