### 2024-05-30 68e13dd
removed refernces to `matplotlib`; fixed error catching
