### 2024-05-30 #6691 @csoneson
👋🏻 @flashton2003 - are you still interested in reviewing this submission for JOSS?

> primerForge: a Python program for identifying primer pairs capable of distinguishing groups of genomes from each other
https://github.com/openjournals/joss-reviews/issues/6691
