### 2024-05-30 20f9056
fix for pypi CI
