### 2024-05-30 #6691 @csoneson
👋🏻 @k-florek, @jrober84 - would you be interested in reviewing this submission for JOSS?

> primerForge: a Python program for identifying primer pairs capable of distinguishing groups of genomes from each other
https://github.com/openjournals/joss-reviews/issues/6691

We perform our checklist-driven reviews on GitHub. More information about the process can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Thanks in advance!
