### 2024-05-31 #6691 @flashton2003
Sure, that’s fine.

On Thu, 30 May 2024 at 22:13, Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @k-florek <https://github.com/k-florek>, @jrober84
> <https://github.com/jrober84> - would you be interested in reviewing this
> submission for JOSS?
>
> primerForge: a Python program for identifying primer pairs capable of
> distinguishing groups of genomes from each other
> #6691 <https://github.com/openjournals/joss-reviews/issues/6691>
>
> We perform our checklist-driven reviews on GitHub. More information about
> the process can be found here
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.
>
> Thanks in advance!
>
