### 2024-05-31 #6691 @flashton2003
Is there an explainer of how this process works somewhere?

On Fri, 31 May 2024 at 08:55, The Open Journals editorial robot <
***@***.***> wrote:

> @flashton2003 <https://github.com/flashton2003> added to the reviewers
> list!
>
