### 2024-05-31 #6691 @csoneson
Yes - see [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Let me know if you have any questions!
