### 2024-06-03 f3c3721
added entrypoints to unit tests
