### 2024-06-03 c9fe77f
made unit tests easier to run
