+++ bin/unit_tests/primer_test.py
@@ -1,3 +1,6 @@
+import pathlib, sys
+sys.path.append(str(pathlib.Path(__file__).parent.parent.parent))
+
 from Bio.Seq import Seq
 import primer3, unittest
 from bin.Primer import Primer
