### 2024-06-03 e1834ef
revised `_getNumThreads` to provide a default after a timeout
