### 2024-06-03 d5ca1dd
better status report
