+++ environment.yml
@@ -16,4 +16,3 @@ dependencies:
   - numpy
   - primer3-py>=2.0
   - scipy>=1.10
-  - unittest
