### 2024-06-04 #6691 @csoneson
👋🏻 @mbhall88, @kapsakcj - would one of you be interested in reviewing this submission for JOSS?

> primerForge: a Python program for identifying primer pairs capable of distinguishing groups of genomes from each other
https://github.com/openjournals/joss-reviews/issues/6691

Thanks in advance!
