### 2024-06-04 #6691 @lskatz
Hi, thank you for helping us through the process! We wanted to let you know that @kapsakcj was in our team until three years ago and it might not be appropriate to have him review our work just yet depending on JOSS policy.
