### 2024-06-12 dr-joe-wirth/primerForge#20 @mbhall88
Failed manual installation
This issues is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/6850

When attempting to perform a manual install I get the following error

```
$ conda env create -f environment.yml
Channels:
 - bioconda
 - defaults
 - conda-forge
Platform: linux-64
Solving environment: failed
Channels:
 - bioconda
 - defaults
 - conda-forge
Platform: linux-64
Solving environment: failed

LibMambaUnsatisfiableError: Encountered problems while solving:
  - package biopython-1.81-py39hf860d4a_1 is excluded by strict repo priority
  - package primer3-py-2.0.0-py311hec16e2b_0 requires libgcc-ng >=12, but none of the providers can be installed
```

My conda version is 24.5.0
