### 2024-06-12 dr-joe-wirth/primerForge#21 @mbhall88
How to run tests
This issues is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/6850

In relation to the documentation guideline **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

I see there is a unit tests directory in your `bin/` directory, but I don't see any documented way of running the tests - doesn't seem to be run on the CI so I couldn't figure it out.
