### 2024-06-12 dr-joe-wirth/primerForge#22 @mbhall88
Add a contributing document
This issues is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/6850

Add a [document outlining how others can contribute to this project](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) (or whether they cannot).

This might also be a good place to add how unit tests can be run as per #21
