### 2024-06-12 d80ff46
added instructions for running unit tests.
