### 2024-06-12 dr-joe-wirth/primerForge#20 @dr-joe-wirth
this is interesting. it works for my using version 23.1.0. i will try to look further into this.

in the meantime, can you tell me what your parameter `channel_priority` is set to? mine is set to `flexible`, and it fails if i set it to `strict` (although it is different from your error). you can check/modify using the following commands:

### getting current value
```bash
$ conda config --show channel_priority
```

### changing current value to `flexible`
```bash
$ conda config --set channel_priority flexible
```

### getting additional information about `channel_priority`
```bash
$ conda config --describe channel_priority
# # channel_priority (ChannelPriority)
# #   Accepts values of 'strict', 'flexible', and 'disabled'. The default
# #   value is 'flexible'. With strict channel priority, packages in lower
# #   priority channels are not considered if a package with the same name
# #   appears in a higher priority channel. With flexible channel priority,
# #   the solver may reach into lower priority channels to fulfill
# #   dependencies, rather than raising an unsatisfiable error. With channel
# #   priority disabled, package version takes precedence, and the
# #   configured priority of channels is used only to break ties. In
# #   previous versions of conda, this parameter was configured as either
# #   True or False. True is now an alias to 'flexible'.
# #
# channel_priority: flexible
```

if your priority is `strict` can you please try setting it to `flexible`?
