### 2024-06-12 dr-joe-wirth/primerForge#21 @mbhall88
Out of interest. You say tests may take three hours to run. How come it takes so long?
