### 2024-06-12 dr-joe-wirth/primerForge#21 @dr-joe-wirth
It will run the software first, on a test case of five genomes which takes some time. The tests in `results_test.py` make sure that the results are sound and congruent with the input genomes.
