### 2024-06-13 dr-joe-wirth/primerForge#20 @mbhall88
Mine is set to strict. However, I am not keen to change this as bioconda require you to [set a strict channel priority too](https://bioconda.github.io/#usage). And I suspect most other users will be in the same boat as me....

Though I have figured out the issue. It's the ordering of the channels in the environment file you supply. When I switch it to

```yaml
channels:
  - conda-forge
  - bioconda
  - defaults
```

as in the bioconda setup, it installs as expected.
