### 2024-06-13 707666e
revision allows strict conda channel priority
