### 2024-06-13 dr-joe-wirth/primerForge#20 @dr-joe-wirth
thanks for this information. i have revised the file to be compatible with strict channel priority
