### 2024-06-14 dr-joe-wirth/primerForge#23 @mbhall88
Estimated resource requirement
As part of https://github.com/openjournals/joss-reviews/issues/6850 I have been trying to run primerForge using MTB as the ingroup and M. smegmatis as the outgroup

```
primerForge -i h37rv.fa -u MSmeg.fa -f fasta -n8
```
so far I have been unable to get this to complete as it keeps hitting my job memory limits. The last run I tried allocated 160GB and the job failed due to out of memory problems... This seems very high. Is this expected? If so, this should probably be documented somewhere. Or am I doing something wrong here? There isn't an example usage (see #2) so I am just basing my executation on the CLI help menu
