### 2024-06-14 dr-joe-wirth/primerForge#23 @mbhall88
This is the MTB genome https://www.ncbi.nlm.nih.gov/nuccore/NC_000962.3 and this is the M. smegmatis genome https://www.ncbi.nlm.nih.gov/nuccore/NC_008596.1

I am using the bioconda installation of primerForge
