### 2024-06-14 dr-joe-wirth/primerForge#24 @mbhall88
Relation to swga
In relation to https://github.com/openjournals/joss-reviews/issues/6850

One of the reviewing criteria for the paper is **State of the field**: Do the authors describe how this software compares to other commonly-used packages?

Might be worth describing how primerForge relates to https://github.com/songlab-cal/swga2. i.e., does it have more customisation, is it faster etc.
