### 2024-06-17 a58072b
installation instructions now compatible with a strict channel priority
