### 2024-06-17 dr-joe-wirth/primerForge#23 @dr-joe-wirth
Can you please rerun using the `--debug` flag and then share the `.log` file with me? Also, how do you determine how much memory was being allocated?
