### 2024-06-17 dr-joe-wirth/primerForge#23 @mbhall88
Okay, running now. Will share the log when it's done.

Well I requested 160GB of memory from slurm and then the job failed with 'out of memory'
