### 2024-06-18 dr-joe-wirth/primerForge#23 @mbhall88
Just failed with 200GB (max RSS 208658340KB)

```
INFO:bin.main:version:                          1.1.1
INFO:bin.main:ingroup:                          /data/scratch/projects/punim1703/WGA/data/references/h37rv.fa
INFO:bin.main:outgroup:                         /data/scratch/projects/punim1703/WGA/data/references/MSmeg.fa
INFO:bin.main:results filename:                 /data/scratch/projects/punim1703/tbvcf/tmp/primerForge/results.tsv
INFO:bin.main:file format:                      fasta
INFO:bin.main:min kmer len:                     16
INFO:bin.main:max kmer len:                     20
INFO:bin.main:min % G+C                         40.0
INFO:bin.main:max % G+C                         60.0
INFO:bin.main:min Tm:                           55.0
INFO:bin.main:max Tm:                           68.0
INFO:bin.main:max Tm difference:                5.0
INFO:bin.main:min PCR size:                     120
INFO:bin.main:max PCR size:                     2400
INFO:bin.main:disallowed outgroup PCR sizes:    120-2400
INFO:bin.main:num threads:                      8
INFO:__getCandidates:identifying kmers suitable for use as primers in all 1 ingroup genome sequences
INFO:_getAllCandidateKmers:    getting shared ingroup kmers that appear once in each genome
DEBUG:__getSharedKmers:        18539592 shared kmers after processing h37rv.fa
INFO:_getAllCandidateKmers:    done 00:06:54.69
INFO:_getAllCandidateKmers:dumping shared kmers to '_pickles/sharedKmers.p'
INFO:_getAllCandidateKmers:done 00:01:09.11
INFO:_getAllCandidateKmers:    evaluating kmers
INFO:_getAllCandidateKmers:    done 00:02:14.37
INFO:_getAllCandidateKmers:    identified 843387 candidate kmers
INFO:__getCandidates:done 00:10:20.55
INFO:__getCandidates:dumping candidate kmers to '_pickles/candidates.p'
INFO:__getCandidates:done 00:00:04.04
INFO:__getUnfilteredPairs:identifying pairs of primers found in all ingroup sequences
```
