### 2024-06-27 d8c91f5
all candidate primers can be stored on the plus strand
