### 2024-06-27 9f5deee
removing primers that are substrings of other primers
