### 2024-06-27 875d5f5
major overhaul of candidate pair retrieval
