+++ bin/getPrimerPairs.py
@@ -140,12 +140,14 @@ def __getBinPairs(binned:dict[str,dict[int,list[Primer]]], minPrimerLen:int, min
     return out
 
 
-def _formsDimers(fwd:Primer, rev:Primer) -> bool:
+def _formsDimers(fwd:str, rev:str, fwdTm:float, revTm:float) -> bool:
     """evaluates if two primers may form primer dimers
 
     Args:
-        fwd (Primer): the forward primer
-        rev (Primer): the reverse primer
+        fwd (str): the forward primer sequence
+        rev (str): the reverse primer sequence
+        fwdTm (float): the forward primer melting temperature
+        revTm (float): the reverse primer melting temperature
 
     Returns:
         bool: indicates if primer dimer formation is possible
@@ -154,7 +156,7 @@ def _formsDimers(fwd:Primer, rev:Primer) -> bool:
     FIVE_DEGREES = 5
 
     # dimer formation is a concern if the Tm is sufficiently high
-    return primer3.calc_heterodimer_tm(str(fwd), str(rev)) >= (min(fwd.Tm, rev.Tm) - FIVE_DEGREES)
+    return primer3.calc_heterodimer_tm(fwd, rev) >= (min(fwdTm, revTm) - FIVE_DEGREES)
 
 
 def __isPairSuitable(fwd:Primer, rev:Primer, minPcr:int, maxPcr:int, maxTmDiff:float) -> tuple[bool,int]:
@@ -178,32 +180,25 @@ def __isPairSuitable(fwd:Primer, rev:Primer, minPcr:int, maxPcr:int, maxTmDiff:f
     
     # unsuitable if the Tm difference is not within the allowed range
     if not abs(fwd.Tm - rev.Tm) <= maxTmDiff: return False, pcrLen
-
-    # # unsuitable if primer dimers can form
-    # if _formsDimers(fwd, rev): return False, pcrLen
     
     return True, pcrLen
 
 
-def __evaluateOnePair(fwd:Primer, rev:Primer, minPcr:int, maxPcr:int, maxTmDiff:float, binPair:tuple[str,int,int]) -> tuple[Primer,Primer,int,tuple[str,int,int]]:
+def __evaluateOnePair(fwd:str, rev:str, fTm:float, rTm:float, binPair:str) -> str:
     """evaluates a primer pair; designed to work in parallel; only returns if the pair passes evaluation
 
     Args:
-        fwd (Primer): the forward Primer
-        rev (Primer): the reverse Primer
-        minPcr (int): the minimum PCR product size
-        maxPcr (int): the maximum PCR product size
-        maxTmDiff (float): the maximum difference between primer Tm
-        binPair (tuple[str,int,int]): contig, bin1 number, bin2 number
+        fwd (str): the forward Primer
+        rev (str): the reverse Primer
+        fTm (float): the forward melting temperature
+        rTm (float): the reverse melting temperature
+        binPair (str): the bin pair for this pair
 
     Returns:
-        tuple[Primer,Primer,int,int,int]: forward primer, reverse primer, pcr product size, bin pair
+        str: the bin pair for this pair
     """
-    # determine the product size and if the pair is suitable for PCR
-    acceptablePair, pcrLen = __isPairSuitable(fwd, rev, minPcr, maxPcr, maxTmDiff)
-    
     # return acceptable pairs and their pcr product sizes
-    if acceptablePair: return fwd,rev.reverseComplement(),pcrLen,binPair
+    if not _formsDimers(fwd, rev, fTm, rTm): return binPair
 
 
[…]
