### 2024-06-28 070fefb
fixed bug associated with binning kmers
