### 2024-06-28 f684563
trying to improve error traceback
