### 2024-07-01 22f9947
major overhaul to `__getUniqueKmers`; significantly improves runtimes   * using `khmer` for kmer retrieval   * added `__getAllowedKmers`   * added `__extractKmers`
