+++ bin/getCandidateKmers.py
@@ -2,13 +2,81 @@ import os
 from Bio import SeqIO
 from Bio.Seq import Seq
 from bin.Clock import Clock
+from khmer import Countgraph
 from bin.Primer import Primer
 import multiprocessing, primer3
 from Bio.SeqRecord import SeqRecord
 from bin.Parameters import Parameters
 
 # functions
-def __getUniqueKmers(fn:str, frmt:str, minLen:int, maxLen:int, name:str) -> tuple[str,dict[str,dict[Seq,dict[str,tuple[str,int,int,str]]]]]:
+def __getAllowedKmers(fwd:str, rev:str, k:int) -> set[str]:
+    """gets all the kmers that are allowed
+
+    Args:
+        fwd (str): the entire (concatenated) forward sequence
+        rev (str): the entire (concatenated) reverse sequence
+        k (int): the length of the kmers
+
+    Returns:
+        set[str]: a set of the allowed kmers
+    """
+    # helper functions
+    def isOneEndGc(seq:str) -> bool:
+        """ evaluates if one end of a sequence is a GC
+        """
+        GC = {"G", "C", 'g', 'c'}
+        return seq[-1] in GC or seq[0] in GC
+    
+    def appearsOnce(seq:str, kh:Countgraph) -> bool:
+        return kh.get(seq) == 1
+    
+    # create count graphs for both strands
+    fkh = Countgraph(k, 1e7, 1)
+    rkh = Countgraph(k, 1e7, 1)
+    
+    # consume the sequences so kmers can be counted
+    fkh.consume(fwd)
+    rkh.consume(rev)
+    
+    # extract the allowed kmers for both strands
+    out = {x for x in fkh.get_kmers(fwd) if isOneEndGc(x) and appearsOnce(x, fkh)}
+    tmp = {x for x in rkh.get_kmers(rev) if isOneEndGc(x) and appearsOnce(x, rkh)}
+    
+    # remove any kmers shared between strands
+    out.difference_update(tmp)
+    
+    return out
+
+
+def __extractKmers(name:str, contig:str, strand:str, seq:str, k:int, allowed:set[str]) -> dict[str,dict[str,tuple[str,int,int,str]]]:
+    """extracts kmers from a contig's sequence
+
+    Args:
+        name (str): the name of the sequence file
+        contig (str): the name of the contig
+        strand (str): the strand of the sequence
+        seq (str): the dna sequence
+        k (int): the length of the kmer
+        allowed (set[str]): a set of allowed kmers
+
+    Returns:
+        dict[str,dict[str,tuple[str,int,int,str]]]: key=kmer; val=dict: key=name; val=(contig, start, klen, strand)
+    """
+    # create a countgraph and consume the sequence
+    kh = Countgraph(k, 1e7, 1)
+    kh.consume(seq)
+    
+    # extract the forward kmers and build the output
+    if strand == Primer.PLUS:
+        return {kmer:{name: (contig, start, k, strand)} for start,kmer in enumerate(kh.get_kmers(seq)) if kmer in allowed}
+    
+    # reverse kmer start position needs to be relative to the plus strand
+    else:
+        length = len(seq)
+        return {kmer:{name: (contig, (length - start - k), k, strand)} for start,kmer in enumerate(kh.get_kmers(seq)) if kmer in allowed}
+
+
+def __getUniqueKmers(fn:str, frmt:str, minK:int, maxK:int, name:str) -> tuple[str,dict[str,dict[str,dict[str,tuple[str,int,int,str]]]]]:
     """gets a collection of kmers from a genome that are:
         * not repeated anywhere in the genome
         * at least one end is a G or C
[…]
