+++ environment.yml
@@ -13,6 +13,7 @@ dependencies:
   - python=3.11.0
   - pip=23.0.1
   - biopython=1.81
+  - khmer>=2.1
   - numpy
   - primer3-py>=2.0
   - scipy>=1.10
