+++ bin/main.py
@@ -10,7 +10,7 @@ from bin.removeOutgroupPrimers import _removeOutgroupPrimers
 from bin.getPrimerPairs import _getPrimerPairs, _keepOnePairPerBinPair
 
 # global constants
-__version__ = "1.2.2"
+__version__ = "1.2.3"
 __author__ = "Joseph S. Wirth"
 __SHARED_NUM = 0
 __CAND_NUM   = 1
@@ -422,7 +422,6 @@ def main() -> None:
         except Exception as e:
             # save the error message if in debug mode
             params.log.critical(e)
-            params.log.critical(e.__traceback__)
             
             Clock._killWheel()
             
