### 2024-07-01 7dc072d
major overhaul to `__getAllKmersForOneContig`   * significant improvements to runtime using `khmer`   * added `__kmerListToDict`
