### 2024-07-01 8109ed5
fixed bug in `__getAllowedKmers`
