### 2024-07-02 8f5f7b2
updated installer check to include `khmer`
