### 2024-07-02 3fe5ce1
fixed bug in `testK_noPrimerDimers`
