### 2024-07-02 c248d6a
fixed bug in `testK_noPrimerDimers`
