### 2024-07-02 dbc3035
fixed bugs in `__getAllKmersForOneContig`; unit tests passing
