### 2024-07-02 b16c327
removed "keep one pair per binpair" step; this now happens earlier
