### 2024-07-02 7243372
improved kmer retrieval using `khmer`
