### 2024-07-02 ecac8a6
added `khmer` to manuscript
