+++ paper/paper.bib
@@ -1,3 +1,14 @@
+{@article{10.12688/f1000research.6924.1,
+   author = {{Crusoe}, Michael R. and {Alameldin}, Hussien F. and {Awad}, Sherine and {Bucher}, Elmar and {Caldwell}, Adam and {Cartwright}, Reed and {Charbonneau}, Amanda and {Constantinides}, Bede and {Edvenson}, Greg and {Fay}, Scott and {Fenton}, Jacob and {Fenzl}, Thomas and {Fish}, Jordan and {Garcia-Gutierrez}, Leonor and {Garland}, Phillip and {Gluck}, Jonathan and {González}, Iván and {Guermond}, Sarah and {Guo}, Jiarong and {Gupta}, Aditi and {Herr}, Joshua R. and {Howe}, Adina and {Hyer}, Alex and {Härpfer}, Andreas and {Irber}, Luiz and {Kidd}, Rhys and {Lin}, David and {Lippi}, Justin and {Mansour}, Tamer and {McA'Nulty}, Pamela and {McDonald}, Eric and {Mizzi}, Jessica and {Murray}, Kevin D. and {Nahum}, Joshua R. and {Nanlohy}, Kaben and {Nederbragt}, Alexander Johan and {Ortiz-Zuazaga}, Humberto and {Ory}, Jeramia and {Pell}, Jason and {Pepe-Ranney}, Charles and {Russ}, Zachary N and {Schwarz}, Erich and {Scott}, Camille and {Seaman}, Josiah and {Sievert}, Scott and {Simpson}, Jared and {Skennerton}, Connor T. and {Spencer}, James and {Srinivasan}, Ramakrishnan and {Standage}, Daniel and {Stapleton}, James A. and {Stein}, Joe and {Steinman}, Susan R and {Taylor}, Benjamin and {Trimble}, Will and {Wiencko}, Heather L. and {Wright}, Michael and {Wyss}, Brian and {Zhang}, Qingpeng and {zyme}, en and {Brown}, C. Titus},
+   title = "{The khmer software package: enabling efficient nucleotide sequence analysis [version 1; peer review: 2 approved, 1 approved with reservations]}",
+   journal = {F1000Research},
+   volume = {4},
+   year = {2015},
+   issn = {2046-1402},
+   doi = {10.12688/f1000research.6924.1},
+   url = {http://dx.doi.org/10.12688/f1000research.6924.1}
+}
+
 @article{10.1093/nar/gks596,
     author = {{Untergasser}, Andreas and {Cutcutache}, Ioana and {Koressaar}, Triinu and {Ye}, Jian and {Faircloth}, Brant C. and {Remm}, Maido, and {Rozen}, Steven G.},
     title = "{Primer3—new capabilities and interfaces}",
