+++ paper/paper.md
@@ -59,9 +59,10 @@ primers capable of producing PCR products of a specific size in a set of whole
 genome sequences. Optionally, it can also filter those primer pairs and limit
 the output to primer pairs that can be used to distinguish one set of genomes
 from another set of genomes via PCR amplification. `primerForge` relies on the
-`primer3-py` package to evaluate specific characteristics of primer pairs
-including melting temperature, hairpin potential, and dimer formation
-[@10.1093/nar/gks596].
+`khmer` package to extract k-mers from genomic sequences and the `primer3-py`
+package to evaluate specific characteristics of primer pairs including melting
+temperature, hairpin potential, and dimer formation
+[@10.12688/f1000research.6924.1; @10.1093/nar/gks596].
 
 There are many use cases for what `primerForge` offers. One use case would be
 surveillance of an outbreak clone of a particular pathogen. A laboratory could
