### 2024-07-02 b238af2
updated mermaid diagram
