### 2024-07-02 dr-joe-wirth/primerForge#23 @dr-joe-wirth
I believe I have resolved this issue. I have improved performance in many places. When I ran with your inputs, I recorded it using <15gb ram. I am going to close this issue; please feel free to report back if I need to reopen it.
