### 2024-07-02 dr-joe-wirth/primerForge#23 @dr-joe-wirth
note: the conda installation and docker image are not currently up and running for the updated version. please use only the manual installation or the pip installation for the time being.
