### 2024-07-02 dr-joe-wirth/primerForge#26 @mbhall88
python 3.12 pip install error
In relation to https://github.com/openjournals/joss-reviews/issues/6850

When I try and pip install the latest (v1.2.5) version with python 3.12 I get the following error

```
python -m pip install primerforge
Collecting primerforge
  Downloading primerforge-1.2.5-py3-none-any.whl.metadata (6.6 kB)
Collecting biopython==1.81 (from primerforge)
  Downloading biopython-1.81-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (13 kB)
Collecting khmer>=2.1 (from primerforge)
  Downloading khmer-2.1.1.tar.gz (8.5 MB)
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [18 lines of output]
      /tmp/pip-install-00inzozi/khmer_9aa20b283bf746d0849a8cfb44703c4d/versioneer.py:421: SyntaxWarning: invalid escape sequence '\s'
        LONG_VERSION_PY['git'] = '''
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/tmp/pip-install-00inzozi/khmer_9aa20b283bf746d0849a8cfb44703c4d/setup.py", line 168, in <module>
          "define_macros": [("VERSION", versioneer.get_version()), ],
                                        ^^^^^^^^^^^^^^^^^^^^^^^^
        File "/tmp/pip-install-00inzozi/khmer_9aa20b283bf746d0849a8cfb44703c4d/versioneer.py", line 1480, in get_version
          return get_versions()["version"]
                 ^^^^^^^^^^^^^^
        File "/tmp/pip-install-00inzozi/khmer_9aa20b283bf746d0849a8cfb44703c4d/versioneer.py", line 1412, in get_versions
          cfg = get_config_from_root(root)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/tmp/pip-install-00inzozi/khmer_9aa20b283bf746d0849a8cfb44703c4d/versioneer.py", line 342, in get_config_from_root
          parser = configparser.SafeConfigParser()
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      AttributeError: module 'configparser' has no attribute 'SafeConfigParser'. Did you mean: 'RawConfigParser'?
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

I have no such problems with python 3.11
