### 2024-07-03 dr-joe-wirth/primerForge#23 @mbhall88
great. can confirm this completed with 8 threads in 13 mins and ~23GB memory.
