### 2024-07-03 592dba5
now compatible with `khmer` dependency
