+++ docker/Dockerfile
@@ -7,7 +7,7 @@ FROM ubuntu:22.04
 
 # install wget and python
 RUN apt-get update && \
-    apt-get -y install python3.11 python3-pip python3-setuptools python3-dev
+    apt-get -y install python3.11 python3-pip python3-setuptools python3.11-dev
 
 # install primerForge.py
 RUN python3.11 -m pip install primerforge
