### 2024-07-03 dr-joe-wirth/primerForge#26 @dr-joe-wirth
thanks for bringing this to my attention. i think this is because python3.12 removed `SafeConfigParser` and replaced it with `ConfigParser`. `khmer` is a fantastic tool, but i am finding that it is not a good neighbor when it comes to installation. i am going to see if i can replace `khmer` with something friendlier.
