### 2024-07-03 d4f6879
updated python requirements to specify 3.9 to 3.11
