+++ environment.yml
@@ -10,7 +10,7 @@ channels:
   - bioconda
   - defaults
 dependencies:
-  - python=3.11.0
+  - python>=3.9,<3.12
   - pip=23.0.1
   - biopython=1.81
   - numpy
