### 2024-07-08 dr-joe-wirth/primerForge#26 @dr-joe-wirth
using python 3.12 and above is not supported by `khmer`. this will not change unless `khmer` releases a new version.
