### 2024-07-08 dr-joe-wirth/primerForge#24 @dr-joe-wirth
I will have to write up a more detailed comparison and include it in the manuscript. I will close this issue once I have done that. In the meantime, here is what I have observed so far:

### Methods:
* Wall-clock time and memory usage was determined using the `\time` command in bash (ie. `\time <command>`).
* 16 cpus were used for both programs
* i ran both programs with the same resources
* i used default settings with the following exception:
  - melting temperature of 55-68
* input files were always fasta format
* times are reported in `hh:mm:ss` format

### Comparing the [`plasmid_example` dataset](https://github.com/songlab-cal/swga2/tree/master/examples/plasmid_example) included with `swga`:

* `swga` identified 210 primer pairs in 00:03:33 and used 138.58 MB ram.
* `primerForge` identified 3,268 primer pairs in 00:00:07 and used 80.67 MB ram.

### Comparing the [`results_test.py` dataset](https://github.com/dr-joe-wirth/primerForge/blob/592dba5a5b2dfcd9f7cf53a6fc0c00da317ce4c2/bin/unit_tests/results_test.py#L39) included with `primerForge` (the ftp paths for these genomes are specified in lines 39-43 of that file).

* `swga` identified 0 primer pairs (the absence of identified primers resulted in a failure during `step3`) in 00:06:02 and used 2.02 GB ram.
* `primerForge` identified 1,639,772 primer pairs in 01:03:05 and used 29.92 GB ram.
