### 2024-07-10 1caf245
major overhaul of kmer retrieval
