+++ bin/getCandidateKmers.py
@@ -5,17 +5,19 @@ from bin.Clock import Clock
 from khmer import Countgraph
 from bin.Primer import Primer
 import multiprocessing, primer3
+from ahocorasick import Automaton
 from Bio.SeqRecord import SeqRecord
 from bin.Parameters import Parameters
 
 # functions
-def __getAllowedKmers(fwd:str, rev:str, k:int) -> set[str]:
+def __getAllowedKmers(fwd:str, rev:str, k:int, first:bool) -> set[str]:
     """gets all the kmers that are allowed
 
     Args:
         fwd (str): the entire (concatenated) forward sequence
         rev (str): the entire (concatenated) reverse sequence
         k (int): the length of the kmers
+        first (bool): indicates if this is the first genome to be processed
 
     Returns:
         set[str]: a set of the allowed kmers
@@ -38,179 +40,167 @@ def __getAllowedKmers(fwd:str, rev:str, k:int) -> set[str]:
     fkh.consume(fwd)
     rkh.consume(rev)
     
-    # extract the allowed kmers for both strands
+    # extract the allowed kmers for the (+) strand
     out    = {x for x in fkh.get_kmers(fwd) if isOneEndGc(x) and appearsOnce(x, fkh)}
     rKmers = {x for x in rkh.get_kmers(rev) if isOneEndGc(x) and appearsOnce(x, rkh)}
     
-    # keep only the kmers that aren't shared between strands
-    out.symmetric_difference_update(rKmers)
+    # remove the reverse kmers if this is the first genome
+    if first:
+        out.difference_update(rKmers)
+    
+    # otherwise remove kmers that are shared between strands
+    else:
+        out.symmetric_difference_update(rKmers)
     
     return out
 
 
-def __extractKmers(name:str, contig:str, strand:str, seq:str, k:int, allowed:set[str]) -> dict[str,dict[str,tuple[str,int,int,str]]]:
-    """extracts kmers from a contig's sequence
+def __getAllAllowedKmers(params:Parameters) -> Automaton:
+    """gets all the allowed kmers for the entire ingroup
+
+    Args:
+        params (Parameters): a Parameters object
+
+    Returns:
+        Automaton: an Aho-Corasick automaton for substring searching
+    """
+    # initialize variables
+    tmp = dict()
+    rec:SeqRecord
+    args = list()
+    out = Automaton()
+    firstGenome = True
+    curLen = -float('inf')
+    
+    # for each file
+    for fn in params.ingroupFns:
+        # concatenate the forward and reverse sequences
+        fwd = ''
+        rev = ''
+        for rec in SeqIO.parse(fn, params.format):
+            fwd += str(rec.seq)
+            rev += str(rec.seq.reverse_complement())
+        
+        # for each kmer length
+        for k in range(params.minLen, params.maxLen+1):
+            # add the sequences and the length to a list of arguments
+            args.append((fwd, rev, k, firstGenome))
+        
+        # reset first genome boolean
+        firstGenome = False
+    
+    # identify the allowed kmers for each k and each genome in parallel
+    pool = multiprocessing.Pool(params.numThreads)
+    allowed = pool.starmap(__getAllowedKmers, args)
+    pool.close()
+    pool.join()
+    
+    # sort the list by kmer length (small to big)
+    allowed.sort(key=lambda x:len(next(iter(x))))
+    
+    # save the shared kmers for each k
+    while allowed != []:
+        # remove the last item from the list
+        a = allowed.pop()
+        
+        # if the same k, then only keep the shared kmers
+        if len(next(iter(a))) == curLen:
+            tmp[curLen].intersection_update(a)
+        
+        # if the k is less, then make a new entry in the list
+        else:
[…]
