### 2024-07-10 b785610
including new dependency `pyarocorasick`
