+++ environment.yml
@@ -15,6 +15,7 @@ dependencies:
   - biopython=1.81
   - numpy
   - primer3-py>=2.0
+  - pyahocorasick>=2.1
   - scipy>=1.10
   - pip:
       - khmer>=2.1
