### 2024-07-11 d87ad99
major overhaul to how primer binding is detected; siginficant performance gains using Aho-Corasick algorithm
