+++ bin/removeOutgroupPrimers.py
@@ -1,73 +1,62 @@
-import multiprocessing
-from Bio.Seq import Seq
 from bin.Clock import Clock
-from khmer import Countgraph
 from bin.Primer import Primer
 from Bio.SeqRecord import SeqRecord
+from ahocorasick import Automaton
 from bin.Parameters import Parameters
 
 # global constant
 __NULL_PRODUCT = ("NA", 0, ())
 
 
-def __kmerListToDict(kmers:list[tuple[str,int]]) -> dict[str,list[int]]:
-    """coerces a list of tuples to a dictionary of lists
+# functions
+def __getAutomaton(pairs:list[tuple[Primer,Primer]]) -> Automaton:
+    """creates an Aho-Corasick Automaton for substring lookup
 
     Args:
-        kmers (list[tuple[str,int]]): key=kmer; val=start position
+        pairs (list[tuple[Primer,Primer]]): a list of primer pairs
 
     Returns:
-        dict[str,list[int]]: key=kmer; val=list of start positions
+        Automaton: an Aho-Corasick Automaton
+    """
+    # initialize output
+    out = Automaton()
+    
+    # add each kmer in the pairs to the automaton once and make the automaton
+    for kmer in {str(x) for y in pairs for x in y}:
+        out.add_word(kmer, kmer)
+    out.make_automaton()
+
+    return out
+
+
+def __extractKmerData(seq:str, strand:str, kmers:Automaton) -> dict[str, list[int]]:
+    """extracts data from a contig sequence for a collection of kmers
+
+    Args:
+        seq (str): the contig sequence
+        strand (str): the strand of the sequence
+        kmers (Automaton): the collection of kmers as an Aho-Corasick Automaton
+
+    Returns:
+        dict[str, list[int]]: key=kmer; val=list of start positions
     """
     # initialize output
     out = dict()
     
-    # save each start position in a list keyed under the kmer
-    for kmer,start in kmers:
+    # for each kmer in the sequence
+    for end,kmer in kmers.iter(seq):
+        # extract the start position if on the 
+        start = end - len(kmer) + 1
+        
+        # modify the start position if on the (-) strand
+        if strand == Primer.MINUS:
+            start = len(seq) - start - 1
+        
+        # build the output
         out[kmer] = out.get(kmer, list())
         out[kmer].append(start)
     
     return out
-
-
-def __getAllKmersForOneContig(name:str, contig:SeqRecord, minLen:int, maxLen:int) -> tuple[str,str,dict[str,dict[Seq,list[int]]]]:
-    """gets all the kmers and their start positions from a contig
-
-    Args:
-        name (str): the name of the genome
-        contig (SeqRecord): a contig as a SeqRecord object
-        minLen (int): the minimum kmer length
-        maxLen (int): the maximum kmer length
-
-    Returns:
-        tuple[str,str,dict[str,dict[Seq,list[int]]]]: (name, contig name, dict: key=strand; val=dict: key=kmer sequence; val=list of start positions)
-    """
-    # initialize variables
-    out = {Primer.PLUS:  dict(),
-           Primer.MINUS: dict()}
-    
-    # extract the contig sequences
-    fwd = str(contig.seq)
-    rev = str(contig.seq.reverse_complement())
-    
-    # get the contig length
-    contigLen = len(contig)
-
-    # for each kmer length
-    for k in range(minLen,maxLen+1):
-        # create graph objects for evaluating each strand
-        fkh = Countgraph(k, 1e7, 1)
-        rkh = Countgraph(k, 1e7, 1)
-        
-        # get all the forward and reverse kmers and their start positions
[…]
