### 2024-07-11 ad286fd
added unit tests to ci
