### 2024-07-11 b90c092
better handling of non-interactive testing
