### 2024-07-12 73955e4
evaluating peak memory usage in an attempt to resolve failing CI
