+++ .github/workflows/python-ci.yml
@@ -40,4 +40,5 @@ jobs:
       - name: unit tests
         shell: bash -el {0}
         run: |
+          free -h
           \time python -m unittest discover -s ./bin/unit_tests/ -p "*_test.py"
