### 2024-07-12 c538997
additional CI debugging
