### 2024-07-15 ae275c9
removed unused variable
