### 2024-07-15 d658473
modified to only produce artifacts on a failed run
