+++ .github/workflows/python-ci.yml
@@ -42,8 +42,8 @@ jobs:
         run: |
           python -m unittest discover -s ./bin/unit_tests/ -p "*_test.py"
       - name: upload pickles
-        if: always()
-        uses: actions/upload-artifact@v2
+        if: failure()
+        uses: actions/upload-artifact@v4
         with:
           name: pickles
           path: test_dir/_pickles/*.p
