### 2024-07-15 a8672fc
added a retry mechanism to `ResultsTest._downloadOneGenome`
