### 2024-07-16 9c0e3cb
fixed bug caused by chimeric junction kmers
