### 2024-07-16 744d8e1
revised how the first genome is selected; now consistent with first genome in `getCandidateKmers`.
