### 2024-07-16 fa13022
added a mycoplasma dataset:   * now tests on Mycoplasma mycoides by default (smaller genomes yields less resource usage)   * can optionally still use the ecoli dataset when running interactively   * added `ResultsTest._selectDataset`
