### 2024-07-16 a141f1b
checking that CI will fail if a test fails
