### 2024-07-16 723f3ea
added python3 compatibilty note; improved unittest instructions
