### 2024-07-17 #6850 @csoneson
👋🏻 Just checking in again - @dr-joe-wirth, were you able to address the previous issue?
@flashton2003 - could you let us know when you might have a first round of comments for the authors (I see that you generated the checklist but didn't start ticking off the points)?

Thanks everyone 🙂
