### 2024-07-17 #6850 @dr-joe-wirth
yes i recently created a new release that addressed the space complexity and time complexity problems. the pip installation and the manual installation and the docker image all work, but i am trying to resolve an issue with the conda installer that is arising because of an incompatible recipe for `khmer`.

i still need to:

- formally compare primerforge to swga
- add a contributing template
- add an example usage
