### 2024-07-26 bc514fc
Primer equality now works with `str` and `Bio.Seq.Seq`
