+++ environment.yml
@@ -13,6 +13,7 @@ dependencies:
   - python>=3.9,<3.12
   - pip=23.0.1
   - biopython=1.81
+  - ispcr
   - numpy
   - primer3-py>=2.0
   - pyahocorasick>=2.1
