### 2024-07-26 32ee7f3
added new process to validate primer pairs with `isPcr`
