### 2024-07-29 72c82ce
first draft of the example
