### 2024-07-29 1dd34f9
clearer http protocol instructions
