### 2024-07-29 bde8ace
handling unquoted wildcard strings
