### 2024-07-29 cbc6f6b
updated docker build file.
