### 2024-07-29 0dea14f
updated dockerfile for v1.3.1
