+++ bin/main.py
@@ -1,6 +1,5 @@
 import os
 from Bio import SeqIO
-from Bio.Seq import Seq
 from bin.Clock import Clock
 from bin.Primer import Primer
 from Bio.SeqRecord import SeqRecord
