### 2024-07-30 49e0fed
fixed bug in `_validatePrimerPairs`
