### 2024-07-30 fabcbcb
moved pickling inside process
