### 2024-07-30 9c0df4b
improved status messages
