### 2024-07-30 92788f5
better logging of shared kmer retrieval
