+++ setup.py
@@ -9,7 +9,7 @@ if pythonVersion < (3, 9) or pythonVersion >= (3, 12):
 
 setup(
     name='primerforge',
-    version='1.3.1',
+    version='1.3.2',
     author='Joseph S. Wirth',
     packages=find_packages(),
     description='software to identify primers that can be used to distinguish genomes',
