### 2024-07-30 bc85cf3
updated unit test instructions
