+++ .github/CONTRIBUTING.md
@@ -13,5 +13,6 @@ Thank you for showing interest in contributing to `primerForge`! Here are few gu
 * Fix the issue on your new branch
   * Make commits of logical and atomic units
   * Add any new unit tests to help with longetivity. Unit tests have a file extension `_test.py`, rely on the `unittest` framework in Python3, and should be placed in the directory `bin/unit_tests`.
+    - unit tests can be run using the command `python3 -m unittest discover -s ./primerForge/bin/unit_tests/ -p "*_test.py"`
   * If the unit test(s) show any errors on your local computer, then fix these errors.
 * Make a pull request __with the `devel` branch__ when all unit tests show no errors.
