### 2024-07-30 7b0044b
added `isPcr` information to manuscript
