+++ paper/paper.bib
@@ -21,3 +21,16 @@
     doi = {10.1093/nar/gks596},
     url = {https://doi.org/10.1093/nar/gks596}
 }
+
+@article{10.1093/bib/bbs038,
+    author = {{Khun}, Robert M. and {Haussler}, David and {Kent}, James},
+    title = "{The UCSC genome browser and associated tools}",
+    journal = {Briefings in Bioinformatics},
+    volume = {14},
+    number = {2},
+    year = {2012},
+    issn = {1467-5463},
+    doi = {10.1093/bib/bbs038},
+    url = {https://doi.org/10.1093/bib/bbs038}
+}
+}
