### 2024-07-30 72084d4
added command capture to the log file
