### 2024-07-31 dr-joe-wirth/primerForge#26 @mbhall88
Also, given the [base environment for the most recent release of miniconda is now python 3.12](https://github.com/conda-forge/miniforge/releases/tag/24.5.0-0), it is important to note in the install instructions that 3.12 is not support.
