### 2024-07-31 dr-joe-wirth/primerForge#26 @dr-joe-wirth
Yeah it is such a shame that `khmer` is no longer supported and the `conda` installation is incompatible outside of Python 3.6 and 3.7. Currently `conda install` won't work until I can some how fix `khmer`'s installer. The `Countgraph` functionality is golden and nothing else (that I'm aware of) provides a way to count the number of times a kmer appears without iterating across the sequence. I'd use a different tool if it could keep the runtime and memory footprint as low as `khmer` does.

I have set up the `pip` installer so it will fail if an incompatible Python version is being used and to report that in the error message. The environment file for manual install will use an appropriate version of Python. I will double check the readme tomorrow to make sure it explicitly states which versions of Python are compatible.
