### 2024-07-31 dr-joe-wirth/primerForge#26 @mbhall88
Ah what a shame! I assume you can't use Sourmash (which seems to be where khmer evolved to) as you don't want minhash kmers?
