### 2024-07-31 916cc4c
replaced `wget` with `curl`
