### 2024-08-01 f233555
added `swga2` comparison
