+++ paper/paper.md
@@ -71,6 +71,89 @@ surveillance of an outbreak clone of a particular pathogen. A laboratory could
 develop a set of PCR reactions to track the population of this outbreak clone
 which could help inform if the population were to grow, shrink, or migrate.
 
+# Comparing `primerForge` to `swga2`
+Another software package, `swga2`, was developed to choose sets of primers that
+selectively amplify one set of genomes but not in another set of genomes
+[@10.1371/journal.pcbi.1010137]. Given this similarity, it was important to
+compare the performance of `primerForge` to that of `swga2`. To do this, the
+sequence files listed in \autoref{tab:sequences} were used as inputs. The
+default parameters were used except that the melting temperature range was set
+at 55°C to 68°C and only a single processor was used. The primer pairs
+identified by the `swga2` and `primerForge` were evaluated using `isPcr` with
+the following additional parameters: `-tileSize=8`, `-minGood=8`, and
+`-minPerfect=8`. These parameters were necessary because the primers identified
+by `swga2` were too short to use the default value of 15. The results of these
+comparisons are shown in \autoref{tab:comparisons}.
+
+Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
+
+|Dataset|Name|NCBI Accession|Group|
+|:------|:--:|:------------:|:---:|
+|`swga2` plasmid example|pcDNA|not provided|ingroup|
+|`swga2` plasmid example|pLTR|not provided|outgroup|
+|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. KH3J|GCF_003034305.1|ingroup|
+|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. B345/93|GCF_003034275.1|ingroup|
+|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. Gemu Goffa|GCF_003034345.1|ingroup|
+|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. capri str. GM12|GCF_900489555.1|outgroup|
+|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. capri str. 80/93|GCF_018389745.1|outgroup|
+|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. 644-PT8|GCF_001650295.1|ingroup|
+|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. AR-0428|GCF_008727175.1|ingroup|
+|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. FDAARGOS_293|GCF_002208865.2|ingroup|
+|`primerForge` _Escherichia coli_ example|_Escherichia coli_ K12 str. MG1655|GCF_000005845|outgroup|
+|`primerForge` _Escherichia coli_ example|_Salmonella enterica_ subsp. enterica serovar Typhimurium str. LT2|GCF_000006945|outgroup|
+|SARS-CoV-2|Severe acute respiratory syndrome coronavirus 2 isolate SARS-CoV-2/human/USA/MA_MGH_00230/2020|MT520374|ingroup|
+|SARS-CoV-2|Severe acute respiratory syndrome coronavirus 2 isolate SARS-CoV-2/human/USA/MA_MGH_00229/2020|MT520263|ingroup|
+|SARS-CoV-2|Severe acute respiratory syndrome coronavirus 2 isolate SARS-CoV-2/human/USA/MA_MGH_00257/2020|MT520479|ingroup|
[…]
