+++ paper/paper.bib
@@ -33,4 +33,16 @@
     doi = {10.1093/bib/bbs038},
     url = {https://doi.org/10.1093/bib/bbs038}
 }
+
+@article{10.1371/journal.pcbi.1010137,
+  author={{Dwivedi-Yu}, Jane A. and {Oppler}, Zachary J. and {Mitchell}, Matthew W. and {Song}, Yun S. and {Brisson}, Dustin},
+  title="{A fast machine-learning-guided primer design pipeline for selective whole genome amplification}",
+  journal={PLOS Computational Biology},
+  volume={19},
+  number={4},
+  year={2023},
+  issn = {},
+  doi = {10.1371/journal.pcbi.1010137},
+  url = {https://doi.org/10.1371/journal.pcbi.1010137}
+}
 }
