+++ paper/paper.md
@@ -89,33 +89,37 @@ Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 
 |Dataset|Name|NCBI Accession|Group|
 |:------|:--:|:------------:|:---:|
-|`swga2` plasmid example|pcDNA|not provided|ingroup|
-|`swga2` plasmid example|pLTR|not provided|outgroup|
-|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. KH3J|GCF_003034305.1|ingroup|
-|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. B345/93|GCF_003034275.1|ingroup|
-|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. mycoides str. Gemu Goffa|GCF_003034345.1|ingroup|
-|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. capri str. GM12|GCF_900489555.1|outgroup|
-|`primerForge` _Mycoplamsa mycoides_ example|_Mycoplasma mycoides_ subsp. capri str. 80/93|GCF_018389745.1|outgroup|
-|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. 644-PT8|GCF_001650295.1|ingroup|
-|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. AR-0428|GCF_008727175.1|ingroup|
-|`primerForge` _Escherichia coli_ example|_Escherichia coli_ O157 str. FDAARGOS_293|GCF_002208865.2|ingroup|
-|`primerForge` _Escherichia coli_ example|_Escherichia coli_ K12 str. MG1655|GCF_000005845|outgroup|
-|`primerForge` _Escherichia coli_ example|_Salmonella enterica_ subsp. enterica serovar Typhimurium str. LT2|GCF_000006945|outgroup|
+|plasmid<sup>1</sup>|pcDNA|not provided|ingroup|
+|plasmid<sup>1</sup>|pLTR|not provided|outgroup|
+|_M. mycoides_<sup>2</sup>|_Mycoplasma mycoides_ subsp. mycoides str. KH3J|GCF_003034305.1|ingroup|
+|_M. mycoides_<sup>2</sup>|_Mycoplasma mycoides_ subsp. mycoides str. B345/93|GCF_003034275.1|ingroup|
+|_M. mycoides_<sup>2</sup>|_Mycoplasma mycoides_ subsp. mycoides str. Gemu Goffa|GCF_003034345.1|ingroup|
+|_M. mycoides_<sup>2</sup>|_Mycoplasma mycoides_ subsp. capri str. GM12|GCF_900489555.1|outgroup|
+|_M. mycoides_<sup>2</sup>|_Mycoplasma mycoides_ subsp. capri str. 80/93|GCF_018389745.1|outgroup|
+|_E. coli_|_Escherichia coli_ O157 str. 644-PT8|GCF_001650295.1|ingroup|
+|_E. coli_|_Escherichia coli_ O157 str. AR-0428|GCF_008727175.1|ingroup|
+|_E. coli_|_Escherichia coli_ O157 str. FDAARGOS_293|GCF_002208865.2|ingroup|
+|_E. coli_|_Escherichia coli_ K12 str. MG1655|GCF_000005845|outgroup|
+|_E. coli_|_Salmonella enterica_ subsp. enterica serovar Typhimurium str. LT2|GCF_000006945|outgroup|
[…]
