### 2024-08-01 7393974
reformatted footnotes
