### 2024-08-01 b659a41
reformatted footnotes
