### 2024-08-01 fd60377
reformatted footnotes
