### 2024-08-01 3e7b5a9
Revert "trying new table format"
This reverts commit ed865e4b7cf84df9ad8406736f93092c661857c5.
