### 2024-08-01 dr-joe-wirth/primerForge#24 @dr-joe-wirth
@lskatz @jchen232

I have added this information to the manuscript in the `devel` branch. There is something stupid going on with the footnote formatting for the tables, but the content is correct.

Please review the changes and share your feedback with me. Here is a link to the pdf that was generated.

https://github.com/dr-joe-wirth/primerForge/actions/runs/10206663374
