+++ paper/paper.md
@@ -111,7 +111,7 @@ The _M. mycoides_ is provided as an example in the `primerForge` repository.
 
 Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 
-|program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs[^2]|`isPcr`-compatible pairs[^3]|validated pairs[^4]|optimized pairs[^5]|
+|program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs ^[2]|`isPcr`-compatible pairs ^[3]|validated pairs ^[4]|optimized pairs ^[5]|
 |:-----:|:-----:|:-------------:|:------:|:-----------:|:-----------------------:|:--------------:|:--------------:|
 |`swga2`|plasmid|23:21|0.136|94|22|22|11|
 |`primerForge`|plasmid|00:10|0.051|3,210|3,210|3,168|2,934|
@@ -122,15 +122,15 @@ Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 |`swga2`|SARS-CoV-2|10:38|0.141|63|7|0|0|
 |`primerForge`|SARS-CoV-2|00:19|0.122|39|39|15|15|
 
-[^2]: the number of primer pairs identified by the program
+^[2]: the number of primer pairs identified by the program
 
-[^3]: the number of primer pairs that generated PCR products with
+^[3]: the number of primer pairs that generated PCR products with
 `isPcr`
 
-[^4]: the number of primer pairs that produced a PCR product in every
+^[4]: the number of primer pairs that produced a PCR product in every
 ingroup genome and no products in any of the outgroup genomes
 
-[^5]: the number of valid primer pairs that produced exactly one PCR
+^[5]: the number of valid primer pairs that produced exactly one PCR
 product in each ingroup genome
 
 Although many of the primer pairs predicted by `primerForge` were not validated
