### 2024-08-02 2dde7af
caret outside of footnote [1]
