+++ paper/paper.md
@@ -110,7 +110,7 @@ Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 
 Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 
-|program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs ^[2]|`isPcr`-compatible pairs ^[3]|validated pairs ^[4]|optimized pairs ^[5]|
+|program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs ^[2]|`isPcr`-compatible pairs ^[3]|validated pairs ^[4]|optimized pairs ^[the number of valid primer pairs that produced exactly one PCR product in each ingroup genome]|
 |:-----:|:-----:|:-------------:|:------:|:-----------:|:-----------------------:|:--------------:|:--------------:|
 |`swga2`|plasmid|23:21|0.136|94|22|22|11|
 |`primerForge`|plasmid|00:10|0.051|3,210|3,210|3,168|2,934|
@@ -128,9 +128,6 @@ Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 
 ^[4]: the number of primer pairs that produced a PCR product in every
 ingroup genome and no products in any of the outgroup genomes
-
-^[5]: the number of valid primer pairs that produced exactly one PCR
-product in each ingroup genome
 
 Although many of the primer pairs predicted by `primerForge` were not validated
 by `isPcr`, this can be attributed to the fact that `primerForge` allows primer
