### 2024-08-02 d45268b
back to caret outside of brackets for all footnotes
