### 2024-08-02 dea2bfe
trying a new method for footnotes in table since JOSS's footnotes are broken.
