### 2024-08-02 e155942
fixing custom footnotes
