+++ paper/paper.md
@@ -106,9 +106,9 @@ Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 |SARS-CoV-2|SARS-CoV-2 isolate human/USA/MA_MGH_00257/2020|MT520479|ingroup|
 |SARS-CoV-2|SARS-CoV-2 isolate Wuhan-Hu-1|NC_045512|outgroup|
 
-\su{1}: The plasmid dataset is provided as an example in the `swga2` repository.
+\su{1}The plasmid dataset is provided as an example in the `swga2` repository.
 
-\su{2}: The _M. mycoides_ is provided as an example in the `primerForge` repository.
+\su{2}The _M. mycoides_ is provided as an example in the `primerForge` repository.
 
 Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 
@@ -123,15 +123,15 @@ Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 |`swga2`|SARS-CoV-2|10:38|0.141|63|7|0|0|
 |`primerForge`|SARS-CoV-2|00:19|0.122|39|39|15|15|
 
-\su{1}: the number of primer pairs identified by the program
+\su{1}The number of primer pairs identified by the program
 
-\su{2}: the number of primer pairs that generated PCR products with
+\su{2}The number of primer pairs that generated PCR products with
 `isPcr`
 
-\su{3}: the number of primer pairs that produced a PCR product in every
+\su{3}The number of primer pairs that produced a PCR product in every
 ingroup genome and no products in any of the outgroup genomes
 
-\su{4}: the number of valid primer pairs that produced exactly one PCR
+\su{4}The number of valid primer pairs that produced exactly one PCR
 product in each ingroup genome
 
 Although many of the primer pairs predicted by `primerForge` were not validated
