### 2024-08-02 13f0a0b
another attempt to fix custom footnotes
