+++ paper/paper.md
@@ -86,6 +86,8 @@ by `swga2` were too short to use the default value of 15. The results of these
 comparisons are shown in \autoref{tab:comparisons}.
 
 Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
+\su{1}The plasmid dataset is provided as an example in the `swga2` repository.
+\su{2}The _M. mycoides_ is provided as an example in the `primerForge` repository.
 
 |Dataset|Name|NCBI Accession|Group|
 |:------|:--:|:------------:|:---:|
@@ -106,11 +108,14 @@ Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 |SARS-CoV-2|SARS-CoV-2 isolate human/USA/MA_MGH_00257/2020|MT520479|ingroup|
 |SARS-CoV-2|SARS-CoV-2 isolate Wuhan-Hu-1|NC_045512|outgroup|
 
-\su{1}The plasmid dataset is provided as an example in the `swga2` repository.
-
-\su{2}The _M. mycoides_ is provided as an example in the `primerForge` repository.
-
-Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
+Table:\label{tab:comparisons} Comparing `swga2` to `primerForge`
+\su{1}The number of primer pairs identified by the program
+\su{2}The number of primer pairs that generated PCR products with
+`isPcr`
+\su{3}The number of primer pairs that produced a PCR product in every
+ingroup genome and no products in any of the outgroup genomes
+\su{4}The number of valid primer pairs that produced exactly one PCR
+product in each ingroup genome
 
 |program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs\su{1}|`isPcr`-compatible pairs\su{2}|validated pairs\su{3}|optimized pairs\su{4}|
 |:-----:|:-----:|:-------------:|:------:|:-----------:|:-----------------------:|:--------------:|:--------------:|
@@ -123,16 +128,7 @@ Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 |`swga2`|SARS-CoV-2|10:38|0.141|63|7|0|0|
 |`primerForge`|SARS-CoV-2|00:19|0.122|39|39|15|15|
 
-\su{1}The number of primer pairs identified by the program
 
-\su{2}The number of primer pairs that generated PCR products with
-`isPcr`
-
-\su{3}The number of primer pairs that produced a PCR product in every
-ingroup genome and no products in any of the outgroup genomes
-
-\su{4}The number of valid primer pairs that produced exactly one PCR
-product in each ingroup genome
 
 Although many of the primer pairs predicted by `primerForge` were not validated
 by `isPcr`, this can be attributed to the fact that `primerForge` allows primer
