### 2024-08-02 f0a9601
one last attempt at custom footnotes
