### 2024-08-02 c47c318
trying different method for superscripted footnotes
