### 2024-08-02 1d0d733
grab latest devel; update small typesetting
