### 2024-08-02 b17f901
Footnote madness (#28)
* trying caret outside of bracket

* update to upload-artifact_v4 and also name the workflow

* carets outside of brackets

* remove the caret for [1]

* caret outside of footnote [1]

* all on one line

* turn [5] into a long citation

* back to caret outside of brackets for all footnotes

* let's go nuts

* carets outside of brackets

* grab latest devel; update small typesetting
