### 2024-08-02 8ea8e2e
grammar; added additional footnote to table 1
