+++ paper/paper.md
@@ -87,15 +87,15 @@ comparisons are shown in \autoref{tab:comparisons}.
 
 Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 
-|Dataset|Name|NCBI Accession|Group|
+|Dataset|Name|NCBI Accession|Group^1^|
 |:------|:--:|:------------:|:---:|
-|plasmid^1^|pcDNA|not provided|ingroup|
-|plasmid^1^|pLTR|not provided|outgroup|
-|_M. mycoides_^2^|_Mycoplasma mycoides_ subsp. mycoides str. KH3J|GCF_003034305.1|ingroup|
-|_M. mycoides_^2^|_Mycoplasma mycoides_ subsp. mycoides str. B345/93|GCF_003034275.1|ingroup|
-|_M. mycoides_^2^|_Mycoplasma mycoides_ subsp. mycoides str. Gemu Goffa|GCF_003034345.1|ingroup|
-|_M. mycoides_^2^|_Mycoplasma mycoides_ subsp. capri str. GM12|GCF_900489555.1|outgroup|
-|_M. mycoides_^2^|_Mycoplasma mycoides_ subsp. capri str. 80/93|GCF_018389745.1|outgroup|
+|plasmid^2^|pcDNA|not provided|ingroup|
+|plasmid^2^|pLTR|not provided|outgroup|
+|_M. mycoides_^3^|_Mycoplasma mycoides_ subsp. mycoides str. KH3J|GCF_003034305.1|ingroup|
+|_M. mycoides_^3^|_Mycoplasma mycoides_ subsp. mycoides str. B345/93|GCF_003034275.1|ingroup|
+|_M. mycoides_^3^|_Mycoplasma mycoides_ subsp. mycoides str. Gemu Goffa|GCF_003034345.1|ingroup|
+|_M. mycoides_^3^|_Mycoplasma mycoides_ subsp. capri str. GM12|GCF_900489555.1|outgroup|
+|_M. mycoides_^3^|_Mycoplasma mycoides_ subsp. capri str. 80/93|GCF_018389745.1|outgroup|
 |_E. coli_|_Escherichia coli_ O157 str. 644-PT8|GCF_001650295.1|ingroup|
 |_E. coli_|_Escherichia coli_ O157 str. AR-0428|GCF_008727175.1|ingroup|
 |_E. coli_|_Escherichia coli_ O157 str. FDAARGOS_293|GCF_002208865.2|ingroup|
@@ -106,12 +106,17 @@ Table: \label{tab:sequences} Datasets used to compare `primerForge` to `swga2`.
 |SARS-CoV-2|SARS-CoV-2 isolate human/USA/MA_MGH_00257/2020|MT520479|ingroup|
 |SARS-CoV-2|SARS-CoV-2 isolate Wuhan-Hu-1|NC_045512|outgroup|
 
-^1^ The plasmid dataset is provided as an example in the `swga2` repository.  
-^2^ The _M. mycoides_ is provided as an example in the `primerForge` repository.
+^1^ The group indicates which set of genomes should be distinguished from each
+other within each dataset. The "ingroup" indicates genomes that should produce
+PCR products when amplified with the primer pairs identified by the software.
+The "outgroup" indicates genomes that should that should not produce PCR
+products, or should produce PCR product sizes outside a user-specified range.  
+^2^ The plasmid dataset is provided as an example in the `swga2` repository.  
+^3^ The _M. mycoides_ is provided as an example in the `primerForge` repository.
 
 Table: \label{tab:comparisons} Comparing `swga2` to `primerForge`
 
-|program|dataset|runtime (mm:ss)|RAM (Gb)|primer pairs^1^|`isPcr`-compatible pairs^2^|validated pairs^3^|optimized pairs^4^|
[…]
