### 2024-08-02 dcaca2f
added more context for interpretting comparisons; grammar
