### 2024-08-02 #6850 @dr-joe-wirth
Hello @csoneson @mbhall88 @flashton2003

I believe I have addressed all the reviewer feedback that I have received. Here is a list (in no particular order) of some of the things that have been done since the review started.

- A comparison of `primerForge` to `swga2` has been added to the manuscript.
- An example has been provided in the README.
- A CONTRIBUTING document has been added and is referenced in the README
- All available installation options should work and are documented in the README.
- The compatible versions of python are clearly stated in the README
- The `pip` installation has been designed to fail and raise an informative error message if an incompatible version of python is being used
- The high memory usage and runtimes have been mitigated; the libraries used to do this are cited in the manuscript.
- A validation step using _in silico_ PCR has been added; the software used to do this is cited in the manuscript.
- Instructions for running unit tests have been added to the README
- issue templates have been added for both bug reports and feature requests

If there are any additional issues that I need to address, please let me know and/or create issues on the repo.
