### 2024-08-04 #6850 @dr-joe-wirth
> Great work @dr-joe-wirth. All boxes ticked for me now! Thank you for the thorough treatment of the comments.

Thanks for the insightful comments!

> One tiny thing, columns 1 and 3 in table 2 in the manuscript seem to be overlapping.

There is something going on with the pdf generator that joss uses. We are looking into ways to fix it.
