### 2024-08-06 #6850 @dr-joe-wirth
@csoneson is there a type-setting process or some other opportunity to manually format the tables? Some of the columns in the table have been poorly formatted by the inara tool that is used to automatically create a pdf.
