### 2024-08-07 #6850 @csoneson
Hi @dr-joe-wirth - I think you can manually control the relative widths of the columns in the table by changing the number of dashes in the line below the header line in the table, i.e., the
```
|:-----------:|:-----:|:-------------:|:------:|
```
part. You can add some more dashes to the columns that you want to give more space, and remove some from the columns that need less.
