### 2024-08-07 6ea8a30
attempting to fix column width of tables
