### 2024-08-07 635a2be
adjusted table column width
