### 2024-08-07 #6850 @dr-joe-wirth
oh thanks so much @csoneson! i believe i have made the tables a lot more legible.
