### 2024-08-12 c3a3549
added equality method
