### 2024-08-12 3383e06
testing equality method
