### 2024-08-12 a3b7de4
moved pickle numbers into class definition
