+++ bin/main.py
@@ -12,13 +12,8 @@ from bin.removeOutgroupPrimers import _removeOutgroupPrimers
 # global constants
 __version__ = "1.3.3"
 __author__ = "Joseph S. Wirth"
-__SHARED_NUM = 0
-__CAND_NUM   = 1
-__PAIR_1_NUM = 2
-__PAIR_2_NUM = 3
-__PAIR_3_NUM = 4
 
-
+# functions
 def __getCheckpoint(params:Parameters) -> tuple[bool,bool,bool,bool,bool]:
     """gets booleans for checkpointing
 
@@ -29,11 +24,11 @@ def __getCheckpoint(params:Parameters) -> tuple[bool,bool,bool,bool,bool]:
         tuple[bool,bool,bool,bool,bool,bool]: each boolean indicates if the pickle exists: shared kmers; candidate kmers; unfiltered pairs; filtered pairs; validated pairs
     """
     # determine if each pickle exists
-    sharedExists = os.path.exists(params.pickles[__SHARED_NUM])
-    candidExists = os.path.exists(params.pickles[__CAND_NUM])
-    unfiltExists = os.path.exists(params.pickles[__PAIR_1_NUM])
-    filterExists = os.path.exists(params.pickles[__PAIR_2_NUM])
-    validsExists = os.path.exists(params.pickles[__PAIR_3_NUM])
+    sharedExists = os.path.exists(params.pickles[Parameters._SHARED])
+    candidExists = os.path.exists(params.pickles[Parameters._CAND])
+    unfiltExists = os.path.exists(params.pickles[Parameters._PAIR_1])
+    filterExists = os.path.exists(params.pickles[Parameters._PAIR_2])
+    validsExists = os.path.exists(params.pickles[Parameters._PAIR_3])
     
     return sharedExists, candidExists, unfiltExists, filterExists, validsExists
 
@@ -124,7 +119,7 @@ def __getUnfilteredPairs(params:Parameters, candidateKmers:dict[str,dict[str,lis
     params.log.info(f"{MSG_2A}{len(pairs)}{MSG_2B}")
     
     # dump the pairs to file
-    params.dumpObj(pairs, params.pickles[__PAIR_1_NUM], "unfiltered pairs", prefix=" "*4)
+    params.dumpObj(pairs, params.pickles[Parameters._PAIR_1], "unfiltered pairs", prefix=" "*4)
     
     return pairs
 
@@ -165,7 +160,7 @@ def __removeOutgroup(params:Parameters, pairs:dict[tuple[Primer,Primer],dict[str
         params.log.info(f"{GAP}{MSG_1A}{startingNum - finalNum}{MSG_1B}{finalNum}{MSG_1C}")
         
         # dump the results
-        params.dumpObj(pairs, params.pickles[__PAIR_2_NUM], "filtered pairs", prefix=GAP)
+        params.dumpObj(pairs, params.pickles[Parameters._PAIR_2], "filtered pairs", prefix=GAP)
 
 
 def __validatePairs(params:Parameters, pairs:dict[tuple[Primer,Primer],dict[str,tuple[str,int,tuple[str,int,int]]]]) -> None:
@@ -196,7 +191,7 @@ def __validatePairs(params:Parameters, pairs:dict[tuple[Primer,Primer],dict[str,
     params.log.info(f'{GAP}{MSG_1A}{startingNum - finalNum}{MSG_1B}{finalNum}{MSG_1C}')
     
     # dump the results
-    params.dumpObj(pairs, params.pickles[__PAIR_3_NUM], "validated pairs", prefix=GAP)
+    params.dumpObj(pairs, params.pickles[Parameters._PAIR_3], "validated pairs", prefix=GAP)
 
 
[…]
