### 2024-08-12 a32a38c
made equality more stable
