### 2024-08-12 5488a1d
more informative error message
