### 2024-08-13 f6b419b
set minimum primer size to 10bp
