### 2024-08-13 9e3d04f
more stringent `isPcr` parameters
