### 2024-08-13 474eaee
better error messaging
