### 2024-08-14 #6850 @dr-joe-wirth
Repeated the benchmarking using version 1.3.5. Updated the manuscript to reflect this change.
