### 2024-08-20 cb23088
reducing redundant operations when evaluating kmers
