### 2024-08-20 30d0e41
trying out generators instead of argument lists
