### 2024-08-21 1cc1f0e
fixed typehinting problems with Generators
