+++ bin/main.py
@@ -11,7 +11,7 @@ from bin.getCandidateKmers import _getAllCandidateKmers
 from bin.removeOutgroupPrimers import _removeOutgroupPrimers
 
 # global constants
-__version__ = "1.3.5"
+__version__ = "1.3.6"
 __author__ = "Joseph S. Wirth"
 
 # functions
