### 2024-08-26 1dadc47
fixed typo in table 2
