### 2024-08-26 #6850 @dr-joe-wirth
Fixed a typo in Table 2. 61 minutes to 81 minutes for `primerForge` on the _E. coli_ dataset.
