### 2024-08-28 #6850 @csoneson
Thanks @flashton2003 - could you clarify

> All fine by me.

There are some unticked boxes in your checklist above, and you had some concerns that required attention some time ago (https://github.com/openjournals/joss-reviews/issues/6850#issuecomment-2245409672). Could you elaborate a bit more on whether there are issues that need the authors' attention? Thanks!
