### 2024-08-29 #6850 @dr-joe-wirth
@csoneson it occurred to me that I have updated the version since the paper. The manuscript discusses [v1.3.5](https://github.com/dr-joe-wirth/primerForge/tree/v1.3.5) but `primerForge` is currently at v1.3.6.

If you need to evaluate v1.3.5, then here are some instructions to help you out:

## installing v1.3.5
since it isn't the latest version you will need to modify the installation instructions as follows:

### `pip` installation
```bash
pip install primerforge==1.3.5
conda install ispcr
```

### manual installation
```bash
git clone --branch v1.3.5 https://github.com/dr-joe-wirth/primerForge.git
conda env create -f ./primerForge/environment.yml
conda activate primerforge
```
