### 2024-08-29 2905d13
parallelized isPcr calls (#30)
