+++ bin/validatePrimers.py
@@ -1,9 +1,10 @@
-import os, subprocess
 from Bio import SeqIO
 from bin.Clock import Clock
+from typing import Generator
 from bin.Primer import Primer
 from Bio.SeqRecord import SeqRecord
 from bin.Parameters import Parameters
+import multiprocessing, os, subprocess
 
 
 # global constants
@@ -38,38 +39,45 @@ def __makeFastaFile(params:Parameters) -> None:
                 SeqIO.write(SeqRecord(contig.seq, contig.id, '', ''), fh, OUT_FORMAT)
 
 
-def __makeQueryFile(params:Parameters, pairs:dict[tuple[Primer,Primer],dict[str,tuple[str,int,tuple[str,int,int]]]]) -> None:
-    """makes the query file for 
+def __makeQueryString(pairs:list[tuple[Primer,Primer]]) -> str:
+    """makes the query string from a list of primer pairs
 
     Args:
-        params (Parameters): a Parameters object
-        pairs (dict[tuple[Primer,Primer],dict[str,tuple[str,int,tuple[str,int,int]]]]): key=Primer pair; dict:key=genome name; val=tuple: contig, pcr product size, bin pair (contig, num1, num2)
+        pairs (list[tuple[Primer,Primer]]): a list of primer pairs
+    
+    Returns:
+        str: the query string
     """
-    # open the file
-    with open(params.queryFn, 'w') as fh:
-        for pair in pairs.keys():
-            # recast the pair as a strings
-            fwd,rev = map(str, pair)
-            
-            # the "key" column will be the pair separated by a pipe
-            key = __KEY_SEP.join((fwd,rev))
-            
-            # write the data for each pair to file; one per line
-            fh.write(__ROW_SEP.join((key,fwd,rev)) + "\n")
+    # initialize the output
+    out = ""
+    
+    # for each pair
+    for pair in pairs:
+        # recast the pair as a strings
+        fwd,rev = map(str, pair)
+        
+        # the "key" column will be the pair separated by a pipe
+        key = __KEY_SEP.join((fwd,rev))
+        
+        # save the data for each pair; one per line
+        out += __ROW_SEP.join((key,fwd,rev)) + "\n"
+    
+    return out
 
 
-def __runPcr(params:Parameters) -> dict[tuple[str,str],set[tuple[str,int]]]:
-    """runs isPcr on the primer pairs
+def __runOnePcr(allContigsFna:str, query:str) -> dict[tuple[str,str],set[tuple[str,int]]]:
+    """runs isPcr on a properly formatted query string
 
     Args:
-        params (Parameters): a Parameters object
+        allContigsFna (str): the filename containing all the contigs
+        query (str): the query to search
 
     Returns:
-        dict[tuple[str,str],set[tuple[str,int]]]: key=primer pair (as strings); val=pcr product size
+        dict[tuple[str,str],set[tuple[str,int]]]: key=primer pair (as strings); val=(contig, pcr product size)
     """
     # constants
     CMD = 'isPcr'
-    ARGS = ['stdout', '-out=bed', '-minGood=6', '-minPerfect=8']
+    ARGS = ['stdin', 'stdout', '-out=bed', '-minGood=6', '-minPerfect=8']
     TILE = '-tileSize='
     EMPTY = ['']
     
@@ -77,15 +85,17 @@ def __runPcr(params:Parameters) -> dict[tuple[str,str],set[tuple[str,int]]]:
     out = dict()
     
     # build command
-    cmd = [CMD, params.allContigsFna, params.queryFn]
+    cmd = [CMD, allContigsFna]
     cmd.extend(ARGS)
     cmd.append(TILE + str(Parameters._MIN_LEN))
     
-    # run the command
-    results = subprocess.run(cmd, capture_output=True, check=True)
[…]
