### 2024-08-31 #6850 @csoneson
@dr-joe-wirth I have gone through the submission, and overall it looks great. A couple of suggestions for the paper:

* Could you please spell out 'USA' in the author affiliations?
* I'm wondering if it would be worth to add another sentence in the very beginning, explaining the terms 'ingroup' and 'outgroup' in the Summary section, since this is aimed at a very broad audience.
* I stumbled on the sentence on lines 40-41. Should 'in' be removed?

You can then generate a new proof with `@editorialbot generate pdf`.

I will make another post below with the additional next steps.
