### 2024-09-05 6b98dec
revisions to address comment  in openjournals/joss-reviews/issues/6850
