### 2024-09-09 #6850 @csoneson
@dr-joe-wirth Thanks for this. The next steps now are outlined in the post above (https://github.com/openjournals/joss-reviews/issues/6850#issuecomment-2322781947).
