### 2024-09-09 #6850 @dr-joe-wirth
@csoneson I have done these things.
* Double check authors and affiliations (including ORCIDs)
  - confirmed
* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
  - v1.3.5
* Archive the release on Zenodo/figshare/etc and post the DOI here.
  - 10.5281/zenodo.13738173
* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
  - confirmed
* Make sure that the license listed for the archive is the same as the software license.
  - confirmed
