### 2024-09-10 #6850 @dr-joe-wirth
> * It looks like the Zenodo archive was not generated from the GitHub release, could that be? The version in Zenodo is "v1" (not "v1.3.5"), and the files are not listed. Could you look into this (see [here](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) for instructions)?
> * Do you want to archive v1.3.5, or the current v1.3.6 (does the latter include changes resulting from the review)?

Hey @csoneson I think I misunderstood what needed to be archived. The archive I created is storing the .tar.gz file for [the v1.3.5 release](https://github.com/dr-joe-wirth/primerForge/releases/tag/v1.3.5) because that was the version used in the manuscript. I was looking at using the Zenodo integration, but it will not archive a repo until a release is created. I am preparing v1.3.7, but I was unsure what would be preferred and ultimately decided to archive the release in the paper instead of the repo itself. If you prefer, I can link the primerForge repo to Zenodo and then share a link once I create a new release.

> * The title of the Zenodo archive is not the same as that of the paper.

What should the name of the Zenodo archive be? I named it "primerForge v1.3.5" but should it match the title of the repo or the title of the manuscript?

> * The author list is almost identical, but middle initials are not indicated in the Zenodo archive.

I will fix this once I know how you prefer me to archive the software.
