### 2024-09-10 4dd9d9c
implemented a sort method before writing results to file (#31)
