+++ bin/sortPrimerPairs.py
@@ -1,0 +1,139 @@
+import os, numpy, primer3
+from bin.Primer import Primer
+from bin.Parameters import Parameters
+
+
+def __getHomopolymerTemps(pair:tuple[Primer,Primer]) -> float:
+    """returns the sum of the homodimer melting temps for a primer pair
+
+    Args:
+        pair (tuple[Primer,Primer]): a pair of primers
+
+    Returns:
+        float: the sum of both primer's homodimer melting temperatures
+    """
+    # parse the pair as strings
+    fwd,rev = map(str, pair)
+    
+    # calculate dimer tms
+    fDimer = primer3.calc_homodimer_tm(fwd)
+    rDimer = primer3.calc_homodimer_tm(rev)
+    
+    # return the sum of these temps
+    return fDimer + rDimer
+
+
+def __getHairpinTemps(pair:tuple[Primer,Primer]) -> float:
+    """gets the sume of hairpin melting tempertaures for a primer pair
+
+    Args:
+        pair (tuple[Primer,Primer]): a pair of primer
+
+    Returns:
+        float: the sum of the hairpin melting tempertures
+    """
+    # parse the pair as strings
+    fwd,rev = map(str, pair)
+
+    # calculate hairpin tms
+    fHairpin = primer3.calc_hairpin_tm(fwd)
+    rHairpin = primer3.calc_hairpin_tm(rev)
+    
+    # return the sum of these temps
+    return fHairpin + rHairpin
+
+
+def __getHeteroDimerTemp(pair:tuple[Primer,Primer]) -> float:
+    """gets the heterodimer melting temperature for a pair of primers
+
+    Args:
+        pair (tuple[Primer,Primer]): a pair of primers
+
+    Returns:
+        float: the heterodimer melting temperature
+    """
+    
+    # parse the pair as strings
+    fwd,rev = map(str, pair)
+
+    return primer3.calc_heterodimer_tm(fwd, rev)
+
+
+def __getGcDiff(pair:tuple[Primer,Primer]) -> float:
+    """finds the G+C content (mol %) difference between a pair of primers
+
+    Args:
+        pair (tuple[Primer,Primer]): a pair of primers
+
+    Returns:
+        float: the difference in G+C content (mol %)
+    """
+    # parse the pair
+    fwd,rev = pair
+    
+    return abs(fwd.gcPer - rev.gcPer)
+
+
+def __getTmDiff(pair:tuple[Primer,Primer]) -> float:
+    """finds the melting temperature difference between a pair of primers
+
+    Args:
+        pair (tuple[Primer,Primer]): a pair of primers
+
+    Returns:
+        float: the difference between melting temperatures
+    """
+    # parse the pair
+    fwd,rev = pair
+
+    return abs(fwd.Tm - rev.Tm)
+
+
+def __getMedianIngroupProductSize(metadata:dict[str,tuple[str,int,tuple[str,int,int]]], ingroup:list[str]) -> float:
+    """calculates the median product size for the ingroup genomes
+
+    Args:
+        metadata (dict[str,tuple[str,int,tuple[str,int,int]]]): the metadata for a primer pair
+
+    Returns:
+        float: the median PCR product size for the ingroup genomes
+    """
+    # determine the genome names from the ingroup file names
+    ingroup = set(map(os.path.basename, ingroup))
+    
+    # extract the sizes from the metadata for the pair
+    sizes = [size for name,(contig,size,binpair) in metadata.items() if name in ingroup]
+    
+    # negate the median so that pairs will be sorted largest to smallest
+    return -numpy.median(sizes)
+
+
[…]
