### 2024-09-10 de650ca
added tests to ensure sort method is functioning properly (#31)
