### 2024-09-11 c8b4584
added sorting for outgroup products; improved ingroup product sorting
