+++ bin/sortPrimerPairs.py
@@ -1,6 +1,85 @@
+from typing import Union
 import os, numpy, primer3
 from bin.Primer import Primer
 from bin.Parameters import Parameters
+
+
+def __parseOutgroupProductSizes(product:Union[str,int]) -> list[int]:
+    """parses the PCR product sizes for an outgroup genome
+
+    Args:
+        product (Union[str,int]): a comma-separated list of ints (str) or a single int (int)
+
+    Returns:
+        list[int]: a list of PCR product sizes
+    """
+    # constant
+    SEP = ","
+    
+    # convert comma-separated list (str) to a list of ints
+    if type(product) is str:
+        out = [int(x) for x in product.split(SEP)]
+    
+    # if it just an int, then make it a list
+    elif type(product) is int:
+        out = [product]
+    
+    return out
+
+
+def __minDiffFromRange(intgr:int, rng:range) -> int:
+    """calculates the minimum distance of an integer from a range
+
+    Args:
+        intgr (int): an integer
+        rng (range): a range
+
+    Returns:
+        int: the minimum distance of the integer from the range
+    """
+    # calculate the differnces from the upper and lower limits
+    lower = abs(intgr - min(rng))
+    upper = abs(intgr - max(rng))
+    
+    # return the minimum distance from the disallowed range
+    return min((lower, upper))
+
+
+def __sortOutgroupProducts(metadata:dict[str,tuple[str,int,tuple[str,int,int]]], outgroup:list[str], disallowed:range) -> tuple[int,int]:
+    """evaluates the outgroup products for a primer pair in order to sort the pair
+
+    Args:
+        metadata (dict[str,tuple[str,int,tuple[str,int,int]]]): the metadata for a primer pair
+        outgroup (list[str]): a list of outgroup filenames
+        disallowed (range): the range that outgroup products are not allowed to be in
+
+    Returns:
+        tuple[int,int]: minimum distance from the range (negated); the number of outgroup products
+    """
+    # initialize variables
+    products = list()
+    
+    # determine the genome names from the ingroup file names
+    outgroup = frozenset(map(os.path.basename, outgroup))
+    
+    # get the outgroup products that are >0bp
+    products = [v[1] for k,v in metadata.items() if k in outgroup and v != 0]
+    
+    # handle situations where there are no products (best case)
+    if products == []:
+        minDiff = float('inf') # largest number possible
+    
+    # otherwise, need to calculate the minimum distance from the disallowed range
+    else:
+        # parse the outgroup products so that we have a list ints
+        products = list(map(__parseOutgroupProductSizes, products))
+        products = [y for x in products for y in x]
+        
+        # get the smallest value distance from the disallowed range
+        minDiff = min(map(lambda x:__minDiffFromRange(x, disallowed), products))
+    
+    # return the minimum distance (negated so larger differences are sorted first) and the number of products
+    return -minDiff, len(products)
 
 
 def __getHomopolymerTemps(pair:tuple[Primer,Primer]) -> float:
@@ -89,7 +168,7 @@ def __getTmDiff(pair:tuple[Primer,Primer]) -> float:
     return abs(fwd.Tm - rev.Tm)
 
 
-def __getMedianIngroupProductSize(metadata:dict[str,tuple[str,int,tuple[str,int,int]]], ingroup:list[str]) -> float:
+def __sortIngroupProducts(metadata:dict[str,tuple[str,int,tuple[str,int,int]]], ingroup:list[str]) -> tuple[float,float]:
     """calculates the median product size for the ingroup genomes
 
     Args:
@@ -99,13 +178,17 @@ def __getMedianIngroupProductSize(metadata:dict[str,tuple[str,int,tuple[str,int,
         float: the median PCR product size for the ingroup genomes
     """
[…]
