### 2024-09-11 3e2131f
revised sort method for outgroup products
