### 2024-09-11 79f726f
only ask for dataset when downloading genomes
