### 2024-09-12 1c1f247
created new class to store PCR product data
