### 2024-09-12 b599e4d
saving primer homodimer/hairpin Tms for O(1) lookup during sorting (#31)
