### 2024-09-12 2f2d628
major overhauls to speed up sorting (#31)   * implemented Product class to replace tuples of values   * saving heterodimer Tm in Product class for O(1) lookup during sorting   * utilizing new Tm fields in the Primer class   * significantly sped up sorting thanks to Product and Primer changes
