### 2024-09-12 d60bda0
checking file writability in a safer way
