### 2024-09-12 4fb2483
reduced redundant operations; better documentation
