### 2024-09-13 #6850 @csoneson
@dr-joe-wirth Sorry for the delay. Are there major changes to version 1.3.7 compared to what was evaluated by the reviewers? I just want to make sure that the paper doesn't suddenly describe things that the reviewers didn't have the chance to comment on. I believe there are also ways of importing previous releases to Zenodo (although I have not tried them myself) - e.g. described in https://github.com/zenodo/zenodo/issues/1463.
