### 2024-09-13 #6850 @csoneson
Regarding the name and author list of the Zenodo archive, they should be the same as those of the paper.
