### 2024-09-13 #6850 @dr-joe-wirth
@csoneson I believe that I have archived the repository the correct way this time. I used that hack to get Zenodo to archive v1.3.5. The doi is [__10.5281/zenodo.13759569__](https://doi.org/10.5281/zenodo.13759569)
