### 2024-09-13 #6850 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
